feat(plasma-core, plasma-ui): Moved Spinner to core
