+++ packages/plasma-core/src/components/Spinner/WhiteSpinnerSvg.tsx
